--- css/about.css
@@ -1,54 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=League+Spartan:wght@100;200;300;400;500;600;700;800;900&display=swap');
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Spartan', sans-serif;
-}
-
-h1 {
-    font-size: 50px;
-    line-height: 64px;
-    color: #222;
-}
-
-h2 {
-    font-size: 46px;
-    line-height: 54px;
-    color: #222;
-}
-
-h4 {
-    font-size: 20px;
-    color: #222;
-}
-
-h6 {
-    font-weight: 800;
-    font-size: 12px;
-}
-
-p {
-    font-size: 16px;
-    color: #465b52;
-}
-
-.section-p1 {
-    padding: 40px 80px;
-}
-
-.section-m1 {
-    padding: 40px 0;
-}
-
-.heading {
-    text-align: center;
-    font-weight: 800;
-    margin-bottom: 2%;
-    text-transform: uppercase;
-}
-
-body {
-    width: 100%;
-}
