--- css/sell.css
@@ -1,124 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=League+Spartan:wght@100;200;300;400;500;600;700;800;900&display=swap');
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Spartan', sans-serif;
-}
-
-h1 {
-    font-size: 50px;
-    line-height: 64px;
-    color: #222;
-}
-
-h2 {
-    font-size: 46px;
-    line-height: 54px;
-    color: #222;
-    font-weight: 600;
-}
-
-h4 {
-    font-size: 20px;
-    color: #222;
-    font-weight: 700;
-}
-
-h6 {
-    font-weight: 800;
-    font-size: 12px;
-}
-
-p {
-    font-size: 16px;
-    color: #465b52;
-}
-
-.section-p1 {
-    padding: 40px 80px;
-}
-
-.section-m1 {
-    padding: 40px 0;
-}
-
-body {
-    width: 100%;
-}
-
-
-#prodetails {
-    display: flex;
-    margin-top: 20px;
-}
-
-#prodetails .single-pro-img {
-    width: 40%;
-    margin-right: 50px;
-}
-
-#main-img {
-    margin-bottom: 5px;
-}
-
-.small-img-group {
-    display: flex;
-    justify-content: space-between;
-}
-
-.small-img-col {
-    flex-basis: 24%;
-    cursor: pointer;
-}
-
-#prodetails .single-pro-details {
-    width: 50%;
-    padding: 30px;
-}
-
-#prodetails .single-pro-details h6 {
-    font-weight: 600;
-}
-
-#prodetails .single-pro-details h4 {
-    padding: 40px 0 20px 0;
-}
-
-#prodetails .single-pro-details h2 {
-    font-size: 26px;
-}
-
-#prodetails .single-pro-details input {
-    width: 12%;
-    height: 47px;
-    padding-left: 10px;
-    font-size: 16px;
-    margin-right: 10px;
-    border-radius: 12px;
-}
-
-#prodetails .single-pro-details input:focus {
-    outline: none;
-}
-
-#prodetails .single-pro-details button {
-    font-size: 14px;
-    font-weight: 700;
-    padding: 12px;
-    background: #4F709C;
-    color: #fff;
-    border: none;
-    border-radius: 8px;
-    width: fit-content;
-    transition: .5s ease;
-}
-
-#prodetails .single-pro-details button:hover {
-    background: #213555;
-}
-
-#prodetails .single-pro-details span {
-    line-height: 25px;
-}
