--- html/sell.html
@@ -1,151 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Scraptel</title>
-
-    <!-- Bootstrap Links -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-
-    <!-- CSS Links -->
-    <link rel="stylesheet" href="../css/sell.css">
-    <link rel="stylesheet" href="../css/index.css">
-
-    <!-- Bootstrap JS Links -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-        crossorigin="anonymous"></script>
-    <!-- <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js"
-                integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p"
-                crossorigin="anonymous"></script>
-                <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js"
-                integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF"
-                crossorigin="anonymous"></script> -->
-
-    <!-- Bootstrap Icons Link -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css">
-
-    <!-- Font Awesome Link -->
-    <script src="https://kit.fontawesome.com/721b3d6eaf.js" crossorigin="anonymous"></script>
-
-</head>
-
-<body>
-    <!-- Navbar -->
-    <section id="header">
-        <a href="#"><img src="/images/bulb.png" class="scraptel-logo" alt="logo" width="10%"></a>
-        <div>
-            <ul id="navbar">
-                <li><a href="/html/index.html">Home</a></li>
-                <li><a href="/html/about.html">About</a></li>
-                <li><a class="active" href="/html/priceList.html">Price List</a></li>
-                <li><a href="/html/signup.html">Sign In/ Sign Up</a></li>
-                <li><a href="/html/cart.html"><i class="fa-duotone fa-dumbbell"></i></a></li>
-            </ul>
-        </div>
-    </section>
-
-    <!-- Product Details -->
-    <section id="prodetails" class="section-p1">
-        <div class="single-pro-img">
-            <img src="../images/phones.png" width="100%" id="main-img" alt="">
-            <div class="small-img-group">
-                <div class="small-img-col">
-                    <img src="/images/phones.png" id="p1" width="100%" class="small-img" alt="">
-                </div>
-                <div class="small-img-col">
-                    <img src="/images/phones.png" id="p2" width="100%" class="small-img" alt="">
-                </div>
-                <div class="small-img-col">
-                    <img src="/images/phones.png" id="p3" width="100%" class="small-img" alt="">
-                </div>
-                <div class="small-img-col">
-                    <img src="/images/phones.png" id="p4" width="100%" class="small-img" alt="">
-                </div>
-            </div>
-        </div>
-        <div class="single-pro-details">
-            <h6>The below price is just an Estimation. This may varry depending on the product size...</h6>
-            <h4 id="title">Fridges</h4>
-            <h2 id="price">$50</h2>
-            <input type="number" min="1" placeholder="Qty:">
-            <button>Add to Sell Cart</button>
-            <h4>Product Details</h4>
-            <span>Lorem ipsum dolor sit amet consectetur adipisicing elit. Voluptatem cum ex itaque iure sapiente fugit, iusto accusamus sequi deleniti excepturi possimus, nulla placeat delectus repellat molestias nesciunt quibusdam tempore rerum cumque suscipit ad ab, modi quasi. Molestiae veniam ullam ipsam?</span>
-        </div>
-    </section>
-
-
-    <!-- Contact -->
-    <section id="contact-us">
-        <h1 class="heading">Any Queries?</h1>
-        <div class="container">
-            <div class="row">
-                <div class="col-md-7">
-                    <h5>Leave A Message</h5>
-                    <div class="mb-3">
-                        <label for="exampleFormControlInput1" class="form-label">Name</label>
-                        <input type="text" class="form-control" id="exampleFormControlInput1"
-                            placeholder="Enter Your Name">
-                    </div>
-                    <div class="mb-3">
-                        <label for="exampleFormControlInput1" class="form-label">Email</label>
-                        <input type="email" class="form-control" id="exampleFormControlInput1"
-                            placeholder="Example: name@example.com">
-                    </div>
-                    <div class="mb-3">
-                        <label for="exampleFormControlTextarea1" class="form-label">Your Message</label>
-                        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3"
-                            placeholder="Message"></textarea>
-                    </div>
-                    <button class="btn btn-primary">Send</button>
-                </div>
-                <div class="col-md-5">
-                    <h5>Contact Us</h5>
-                    <div class="mt-5">
-                        <div class="d-flex">
-                            <i class="bi bi-geo-alt-fill"></i>
-                            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo, corporis.</p>
-                        </div>
-                        <hr>
-                        <div class="d-flex">
-                            <i class="bi bi-telephone-fill"></i>
-                            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo, corporis.</p>
-                        </div>
-                        <hr>
-                        <div class="d-flex">
-                            <i class="bi bi-envelope-fill"></i>
-                            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo, corporis.</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-
-    <!-- Footer -->
-    <section id="footer">
-        <div class="container">
-            <div class="row row-cols-sm-1">
-                <div class="col-md-8">
-                    <p class="footer-text">Copyright 2023 Scraptel</p>
-                </div>
-                <div class="col-md-4 footer-icons">
-                    <label for="">Follow Us: </label>
-                    <i class="bi bi-instagram"></i>
-                    <i class="bi bi-facebook"></i>
-                    <i class="bi bi-twitter"></i>
-                </div>
-            </div>
-        </div>
-    </section>
-
-    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
-    <script src="../javascript/sell.js"></script>
-</body>
-
-</html>
