+++ public/css/priceList.css
@@ -1,0 +1,106 @@
+@import url('https://fonts.googleapis.com/css2?family=League+Spartan:wght@100;200;300;400;500;600;700;800;900&display=swap');
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: 'Spartan', sans-serif;
+}
+
+h1 {
+    font-size: 50px;
+    line-height: 64px;
+    color: #222;
+}
+
+h2 {
+    font-size: 46px;
+    line-height: 54px;
+    color: #222;
+}
+
+h4 {
+    font-size: 20px;
+    color: #222;
+}
+
+h6 {
+    font-weight: 800;
+    font-size: 12px;
+}
+
+p {
+    font-size: 16px;
+    color: #465b52;
+}
+
+.section-p1 {
+    padding: 40px 80px;
+}
+
+.section-m1 {
+    padding: 40px 0;
+}
+
+body {
+    width: 100%;
+}
+
+#product {
+    text-align: center;
+}
+
+#product .pro-container {
+    display: flex;
+    justify-content: space-between;
+    padding: 20px 0;
+    flex-wrap: wrap;
+}
+
+#product .pro {
+    width: 23%;
+    min-width: 250px;
+    padding: 30px 12px;
+    border: 1px solid #cce7d0;
+    border-radius: 20px;
+    cursor: pointer;
+    box-shadow: 20px 20px 30px rgba(0, 0, 0, 0.05);
+    margin: 20px 0;
+    background: #F5EFE7;
+    transition: .2s ease;
+    position: relative;
+}
+
+#product .pro:hover {
+    box-shadow: 20px 20px 30px rgba(0, 0, 0, 0.09);
+}
+
+#product .pro img {
+    width: 100%;
+    border-radius: 20px;
+}
+
+#product .description h3 {
+    text-align: start;
+    font-size: 25px;
+    font-weight: 600;
+}
+
+#product .description h4 {
+    text-align: start;
+    color: #4F709C;
+    padding-top: 7px;
+}
+
+#product .pro .cart {
+    width: 40px;
+    height: 40px;
+    line-height: 40px;
+    border-radius: 50px;
+    color: #4F709C;
+    background-color: #fff;
+    border: 1px solid #4F709C;
+    position: absolute;
+    bottom: 25px;
+    right: 5px;
+}
