+++ public/javascript/signup.js
@@ -12,5 +12,5 @@ loginLink.addEventListener('click' ,()=>{
 // Closing icon work
 
 document.getElementById("icon-close").addEventListener("click", function() {
-    window.open("/html/index.html", "_self");
+    window.open("/", "_self");
 });
