+++ views/priceList.ejs
@@ -1,0 +1,142 @@
+<%- include("partials/header");-%>    
+            
+            <link rel="stylesheet" href="/css/priceList.css">
+            <section id="product" class="section-p1">
+                <h2>Price List</h2>
+                <p>See How You Can Sell Your Goods To Us</p>
+                <div class="pro-container">
+                    <div class="pro">
+                        <img src="/images/deal.png" alt="">
+                        <div class="description">
+                            <h3 class="title">Smart Phones</h3>
+                            <h4 class="price">Rs. 40</h4>
+                        </div>
+                        <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+                    </div>
+                    <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Fridges</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+            <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Smart Phones</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+            <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Smart Phones</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+            <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Smart Phones</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+            <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Smart Phones</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+            <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Smart Phones</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+            <div class="pro">
+                <img src="/images/deal.png" alt="">
+                <div class="description">
+                    <h3 class="title">Smart Phones</h3>
+                    <h4 class="price">Rs. 40</h4>
+                </div>
+                <a href="#"><i class="fa-solid fa-cart-plus cart"></i></a>
+            </div>
+        </div>
+    </section>
+
+     <!-- Contact -->
+    <section id="contact-us">
+        <h1 class="heading">Any Queries?</h1>
+        <div class="container">
+            <div class="row">
+                <div class="col-md-7">
+                    <h5>Leave A Message</h5>
+                    <div class="mb-3">
+                        <label for="exampleFormControlInput1" class="form-label">Name</label>
+                        <input type="text" class="form-control" id="exampleFormControlInput1"
+                            placeholder="Enter Your Name">
+                    </div>
+                    <div class="mb-3">
+                        <label for="exampleFormControlInput1" class="form-label">Email</label>
+                        <input type="email" class="form-control" id="exampleFormControlInput1"
+                            placeholder="Example: name@example.com">
+                    </div>
+                    <div class="mb-3">
+                        <label for="exampleFormControlTextarea1" class="form-label">Your Message</label>
+                        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3"
+                            placeholder="Message"></textarea>
+                    </div>
+                    <button class="btn btn-primary">Send</button>
+                </div>
+                <div class="col-md-5">
+                    <h5>Contact Us</h5>
+                    <div class="mt-5">
+                        <div class="d-flex">
+                            <i class="bi bi-geo-alt-fill"></i>
+                            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo, corporis.</p>
+                        </div>
+                        <hr>
+                        <div class="d-flex">
+                            <i class="bi bi-telephone-fill"></i>
+                            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo, corporis.</p>
+                        </div>
+                        <hr>
+                        <div class="d-flex">
+                            <i class="bi bi-envelope-fill"></i>
+                            <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Quo, corporis.</p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
+
+    <!-- Footer -->
+    <section id="footer">
+        <div class="container">
+            <div class="row row-cols-sm-1">
+                <div class="col-md-8">
+                    <p class="footer-text">Copyright 2023 Scraptel</p>
+                </div>
+                <div class="col-md-4 footer-icons">
+                    <label for="">Follow Us: </label>
+                    <i class="bi bi-instagram"></i>
+                    <i class="bi bi-facebook"></i>
+                    <i class="bi bi-twitter"></i>
+                </div>
+            </div>
+        </div>
+    </section>
+
+    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
+    <script src="/javascript/priceList.js"></script>
+</body>
+</html>
