+++ views/signup.ejs
@@ -40,7 +40,7 @@
         </div>
         <div class="form-box register"> 
             <h2>Sign in</h2>
-            <form action="#">
+            <form action="/">
                 <div class="input-box">
                     <span class="icon">
                         <ion-icon name="person-outline"></ion-icon>
